DFA-529 refactor app
Create controllers for forms and decide (evaluate) pages.

Create routes that use those controllers and routes for the general site.

Move things that aren't routes, views or controllers into the lib folder.

Move email validator into validation.
