+++ src/routes/decide.ts
@@ -1,0 +1,23 @@
+import express from 'express';
+import {
+    designPatterns, decide,
+    privateBeta,
+    requestSubmitted,
+    showRequestForm,
+    submitRequestForm,
+    timescales,
+    userJourneys
+} from "../controllers/decide";
+
+const router = express.Router();
+
+router.get('/', decide);
+router.get('/timescales', timescales);
+router.get('/private-beta', privateBeta);
+router.get('/user-journeys', userJourneys);
+router.get('/design-patterns', designPatterns);
+router.get('/private-beta/request-form', showRequestForm);
+router.post('/private-beta/request-form', submitRequestForm);
+router.get('/private-beta/request-submitted', requestSubmitted);
+
+export default router;
